rename package pypore.io to pypore.i_o to fix #12.
