+++ src/pypore/event_finder.pyx
@@ -23,9 +23,9 @@ from libc.math cimport fabs
 
 from cpython cimport bool
 
-from pypore.io import get_reader_from_filename
+from pypore.i_o import get_reader_from_filename
 
-from pypore.io.abstract_reader cimport AbstractReader
+from pypore.i_o.abstract_reader cimport AbstractReader
 from pypore.strategies.baseline_strategy cimport BaselineStrategy
 from pypore.strategies.adaptive_baseline_strategy import AdaptiveBaselineStrategy
 
@@ -579,8 +579,8 @@ def find_events(data, parameters=Parameters(), h5file=None, save_file_names=None
 
     :param data: List of data to search. Each item in the list can be one of the following:
 
-        #. An already opened reader. A subclass of :py:class:`pypore.io.abstract_reader.AbstractReader`.\
-           For example, a :py:class:`pypore.io.chimera_reader.ChimeraReader`.
+        #. An already opened reader. A subclass of :py:class:`pypore.i_o.abstract_reader.AbstractReader`.\
+           For example, a :py:class:`pypore.i_o.chimera_reader.ChimeraReader`.
         #. A string filename to be opened. The appropriate reader will be chosen based on the file extension.
 
     :param pipe: (Optional) :py:class:`multiprocessing.Pipe` for status updates during the run.\
