+++ src/pypore/i_o/abstract_reader.pxd
@@ -6,7 +6,7 @@ cdef class AbstractReader:
 
         -Subclasses should implement opening full data files.
         -Subclasses should implement lazy loading with\
-            :py:func:`pypore.io.abstract_converter.get_next_blocks`
+            :py:func:`pypore.i_o.abstract_converter.get_next_blocks`
 
     """
