+++ src/pypore/i_o/tests/testDataFiles/chimera_nonoise_1event_2levels.log
