+++ src/pypore/i_o/tests/testDataFiles/chimera_nonoise_2events_1levels.log
