+++ src/pypore/i_o/tests/testDataFiles/smallChimera.mat
